Add MicrofrontendContext as microfrontend store
Before this change, we relied on the reference of a microfrontend
function component that could either exist or not in the TinyFrontendServerStorage
object.

With this new set of changes, we can reuse the reference if it was
already defined server-side instead of having to CSR every time.

If the user visits a page where a mfe is not needed, the mfe will be
added to the store, and can be used later.
On the other hand, if a page where the mfe is needed is the first one
being visited, then it will SSR+CSR.